% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snacks.R
\name{snacks}
\alias{snacks}
\title{snackin bigtime}
\usage{
snacks(grapes, apples, burritos)
}
\arguments{
\item{grapes}{mass of grape in gram}

\item{apples}{number of apples in bucket}

\item{burritos}{how many burros grubbed}
}
\value{
have you ever smelled the 45?
}
\description{
snackin bigtime
}
