% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surf_score.R
\name{surf_score}
\alias{surf_score}
\title{Surf Score
this will find the score of a surfer's wave based on three judging criteria: speed, power, and flow}
\usage{
surf_score(speed, power, flow)
}
\arguments{
\item{speed}{how fast the surfer is hitting turns... mo spray = mo speed}

\item{power}{how hard the surfer is smacking the lip.. ho bra!}

\item{flow}{how stylin this cat looks as they shred}
}
\description{
Surf Score
this will find the score of a surfer's wave based on three judging criteria: speed, power, and flow
}
